lessons/lesson.23/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-debug-toolbar = "^4.4.6"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-yasg = "^1.12.7"
setuptools = "^71.1.0"
